+++ packages/stellar-wallet-snap/src/handlers/cronjob/refreshConfirmationContext/priceRefresher.ts
@@ -82,10 +82,13 @@ export class ConfirmationPriceRefresher implements IConfirmationContextRefresher
 
       const updatedTokenPrices = uniqueAssetCaipIds.reduce<
         ContextWithPrices['tokenPrices']
-      >((acc, assetId) => {
-        acc[assetId] = prices[assetId]?.price.toString() ?? null;
-        return acc;
-      }, {});
+      >(
+        (acc, assetId) => {
+          acc[assetId] = prices[assetId]?.price.toString() ?? null;
+          return acc;
+        },
+        {} as ContextWithPrices['tokenPrices'],
+      );
 
       return {
         result: {