+++ packages/stellar-wallet-snap/CHANGELOG.md
@@ -11,4 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - This package was migrated from [snap-stellar-wallet](https://github.com/MetaMask/snap-stellar-wallet). See the source repository for the original [changelog](https://github.com/MetaMask/snap-stellar-wallet/blob/main/packages/snap/CHANGELOG.md) ([#161](https://github.com/MetaMask/internal-snaps/pull/161))
 
+### Fixed
+
+- Fixed a TypeScript error in the confirmation price refresher reducer ([#165](https://github.com/MetaMask/internal-snaps/pull/165))
+
 [Unreleased]: https://github.com/MetaMask/internal-snaps/